feat: refactor the sensitivty analysis experiment
